Remove close button from cards.
Closing the card from Huboard doesn't provide for commenting or label change.
We comment a ticket anytime it's closed, so this breaks our workflow.

We've had a few accidental clicks, so we'll disabled the button for now.

A better solution would be to prompt the user for comment and allow label changes.